requirements.txt
anyio==3.7.0
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
numpy==1.25.0
pandas==2.0.2
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.3
